+++ .github/workflows/build_test.yml
@@ -18,6 +18,8 @@ jobs:
       run: cargo build --verbose
     - name: Run tests
       run: cargo test --verbose
+    - name: Test Provider
+      run: cargo test --verbose -p rc-tests
     - name: Get hacspec function stats
       run: bash lib/get_func_stats.sh
     - name: Build and test F* compiler
