+++ provider/Cargo.toml
@@ -9,8 +9,7 @@ hacspec-lib = { path = "../lib" }
 hacspec-chacha20poly1305 = { path = "../examples/hacspec-chacha20poly1305" }
 hacspec-chacha20 = { path = "../examples/hacspec-chacha20" }
 hacspec-poly1305 = { path = "../examples/hacspec-poly1305" }
-rand = { version = "0.7", optional = false }
-rand_core = { version = "0.5", optional = false }
+aead = "0.3"
 
 [dev-dependencies]
 evercrypt-provider = { path = "../utils/evercrypt-provider" }
