+++ provider/src/lib.rs
@@ -1,30 +1,14 @@
-use rand::{self, AsByteSliceMut};
-use rand_core::{OsRng, RngCore};
+//! # Hacspec Provider
+//!
+//! This crate provides a high-level API for hacspec implementations in
+//! [examples](https://github.com/hacspec/hacspec/tree/master/examples).
+//! The provider implements the [RustCrypto traits](https://github.com/RustCrypto/traits)
+//! API to be interoperable with other implementations.
+//!
+//! This crate does not have tests on its own but is tested through the
+//! [utils test](https://github.com/hacspec/hacspec/tree/master/utils/rc-tests).
 
 mod provider;
-pub mod traits;
-pub use provider::Chacha20Poly1305Hacspec;
-
-fn clone_into_array<A, T>(slice: &[T]) -> A
-where
-    A: Default + AsMut<[T]>,
-    T: Clone,
-{
-    let mut a = Default::default();
-    A::as_mut(&mut a).clone_from_slice(slice);
-    a
-}
-
-/// Generate a random byte vector of length `len`.
-pub fn get_random_vec(len: usize) -> Vec<u8> {
-    let mut out = vec![0u8; len];
-    OsRng.fill_bytes(out.as_byte_slice_mut());
-    out
-}
-
-/// Generate a random array.
-pub fn get_random_array<A: Default + AsByteSliceMut>() -> A {
-    let mut out = A::default();
-    OsRng.fill_bytes(out.as_byte_slice_mut());
-    out
-}
+pub use provider::{
+    aead, Aead, AeadInPlace, Chacha20Poly1305, Error, Key, NewAead, Nonce, Payload, Tag,
+};
