+++ utils/evercrypt-provider/src/provider.rs
@@ -1,55 +1,95 @@
-use super::*;
-use evercrypt::prelude::*;
-use hacspec_provider::traits::*;
+use aead::{
+    consts::{U0, U12, U16, U32},
+    generic_array::{ArrayLength, GenericArray},
+    Error,
+};
+pub use aead::{Aead, AeadInPlace, NewAead, Payload};
+use evercrypt::aead::{key_gen, nonce_gen, Aead as EvercryptAead, Mode, Nonce as EvercryptNonce};
 
-pub struct Chacha20Poly1305Evercrypt {}
+pub struct Chacha20Poly1305 {
+    aead: Option<EvercryptAead>,
+}
 
-impl Chacha20Poly1305 for Chacha20Poly1305Evercrypt {
-    fn new() -> Self {
-        Self {}
+impl Chacha20Poly1305 {
+    pub fn key_gen() -> Key {
+        Key::clone_from_slice(&key_gen(Mode::Chacha20Poly1305))
     }
-
-    // Nonce and key generation helper.
-    fn key_gen(&self) -> [u8; 32] {
-        clone_into_array(&aead_key_gen(AeadMode::Chacha20Poly1305))
+    pub fn nonce_gen() -> Nonce {
+        Nonce::from(nonce_gen(Mode::Chacha20Poly1305))
     }
-    fn nonce_gen(&self) -> [u8; 12] {
-        aead_nonce_gen(AeadMode::Chacha20Poly1305)
-    }
+}
 
-    // Single-shot encryption/decryption.
-    fn encrypt(
+pub type Key = GenericArray<u8, U32>;
+pub type Nonce = GenericArray<u8, U12>;
+pub type Tag = GenericArray<u8, U16>;
+
+impl NewAead for Chacha20Poly1305 {
+    type KeySize = U32;
+
+    fn new(key: &Key) -> Self {
+        let aead = EvercryptAead::new(Mode::Chacha20Poly1305, key).ok();
+        Self { aead }
+    }
+}
+
+impl AeadInPlace for Chacha20Poly1305 {
+    type NonceSize = U12;
+    type TagSize = U16;
+    type CiphertextOverhead = U0;
+
+    fn encrypt_in_place_detached(
         &self,
-        key: &[u8; 32],
-        nonce: &[u8; 12],
-        aad: &[u8],
-        m: &[u8],
-    ) -> Result<(Vec<u8>, [u8; 16]), Error> {
-        let (ctxt, tag) = match Aead::new(AeadMode::Chacha20Poly1305, key) {
-            Ok(cipher) => match cipher.encrypt(m, nonce, aad) {
+        nonce: &Nonce,
+        associated_data: &[u8],
+        buffer: &mut [u8],
+    ) -> Result<Tag, Error> {
+        let nonce: EvercryptNonce = into_array(&nonce);
+        let (ctxt, tag) = match &self.aead {
+            Some(cipher) => match cipher.encrypt(buffer, &nonce, associated_data) {
                 Ok(r) => r,
-                Err(e) => return Err(Error(format!("Error: {:?}", e))),
+                Err(_) => return Err(Error),
             },
-            Err(e) => return Err(Error(format!("Error: {:?}", e))),
+            None => return Err(Error),
         };
-        Ok((ctxt, tag))
+        buffer
+            .iter_mut()
+            .zip(ctxt.iter())
+            .for_each(|(dst, &src)| *dst = src);
+        Ok(Tag::clone_from_slice(&tag))
     }
-
-    fn decrypt(
+    fn decrypt_in_place_detached(
         &self,
-        key: &[u8; 32],
-        nonce: &[u8; 12],
-        aad: &[u8],
-        c: &[u8],
-        tag: &[u8; 16],
-    ) -> Result<Vec<u8>, Error> {
-        let ptxt = match Aead::new(AeadMode::Chacha20Poly1305, key) {
-            Ok(cipher) => match cipher.decrypt(c, tag, nonce, aad) {
+        nonce: &Nonce,
+        associated_data: &[u8],
+        buffer: &mut [u8],
+        tag: &Tag,
+    ) -> Result<(), Error> {
+        let nonce: EvercryptNonce = into_array(&nonce);
+        let tag: Tag = into_array(&tag);
+        let ptxt = match &self.aead {
+            Some(cipher) => match cipher.decrypt(buffer, &tag, &nonce, associated_data) {
                 Ok(r) => r,
-                Err(e) => return Err(Error(format!("Error: {:?}", e))),
+                Err(_) => return Err(Error),
             },
-            Err(e) => return Err(Error(format!("Error: {:?}", e))),
+            None => return Err(Error),
         };
-        Ok(ptxt)
+        buffer
+            .iter_mut()
+            .zip(ptxt.iter())
+            .for_each(|(dst, &src)| *dst = src);
+        Ok(())
     }
+}
+
+// GenericArray to arrays.
+
+fn into_array<A, T, N>(ga: &GenericArray<T, N>) -> A
+where
+    A: Default + AsMut<[T]>,
+    T: Clone,
+    N: ArrayLength<T>,
+{
+    let mut a = Default::default();
+    A::as_mut(&mut a).clone_from_slice(ga.as_slice());
+    a
 }
