+++ utils/rc-tests/Cargo.toml
@@ -1,18 +1,15 @@
 [package]
-name = "evercrypt-provider"
+name = "rc-tests"
 version = "0.1.0"
 authors = ["Franziskus Kiefer <franziskuskiefer@gmail.com>"]
 edition = "2018"
 
-[features]
-random = []
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-evercrypt = "0.0.3"
+evercrypt-provider = { path = "../evercrypt-provider" }
 hacspec-provider = { path = "../../provider" }
-aead = "0.3"
 
 [dev-dependencies]
 serde_json = "1.0"
 serde = {version = "1.0", features = ["derive"]}
-rand = "0.7"
