+++ utils/rc-tests/src/lib.rs
@@ -1,0 +1,9 @@
+//! # Provider tests
+//!
+//! This is a dummy crate that implements tests (wycheproof) for now on the
+//! [RustCrypto traits](https://github.com/RustCrypto/traits) API for different
+//! providers.
+//!
+//! Tested providers:
+//! * hacspec
+//! * evercrypt
